requirements.txt
langchain-groq
langgraph
langchain-core
mysql-connector-python
ragas
datasets
prettytable
colorama
python-dotenv